Implement IntervalTree using IntrusiveRedBlackTree
