+++ Ryujinx.Common/Collections/IntrusiveRedBlackTree.cs
@@ -6,17 +6,8 @@ namespace Ryujinx.Common.Collections
     /// Tree that provides the ability for O(logN) lookups for keys that exist in the tree, and O(logN) lookups for keys immediately greater than or less than a specified key.
     /// </summary>
     /// <typeparam name="T">Derived node type</typeparam>
-    public class IntrusiveRedBlackTree<T> where T : IntrusiveRedBlackTreeNode<T>, IComparable<T>
+    public class IntrusiveRedBlackTree<T> : IntrusiveRedBlackTreeImpl<T> where T : IntrusiveRedBlackTreeNode<T>, IComparable<T>
     {
-        protected const bool Black = true;
-        protected const bool Red = false;
-        protected T Root = null;
-
-        /// <summary>
-        /// Number of nodes on the tree.
-        /// </summary>
-        public int Count { get; protected set; }
-
         #region Public Methods
 
         /// <summary>
@@ -82,46 +73,6 @@ namespace Ryujinx.Common.Collections
                 }
             }
             return null;
-        }
-
-        /// <summary>
-        /// Finds the node whose key is immediately greater than <paramref name="node"/>.
-        /// </summary>
-        /// <param name="node">Node to find the successor of</param>
-        /// <returns>Successor of <paramref name="node"/></returns>
-        public static T SuccessorOf(T node)
-        {
-            if (node.Right != null)
-            {
-                return Minimum(node.Right);
-            }
-            T parent = node.Parent;
-            while (parent != null && node == parent.Right)
-            {
-                node = parent;
-                parent = parent.Parent;
-            }
-            return parent;
-        }
-
-        /// <summary>
-        /// Finds the node whose key is immediately less than <paramref name="node"/>.
-        /// </summary>
-        /// <param name="node">Node to find the predecessor of</param>
-        /// <returns>Predecessor of <paramref name="node"/></returns>
-        public static T PredecessorOf(T node)
-        {
-            if (node.Left != null)
-            {
-                return Maximum(node.Left);
-            }
-            T parent = node.Parent;
-            while (parent != null && node == parent.Left)
-            {
-                node = parent;
-                parent = parent.Parent;
-            }
-            return parent;
         }
 
         #endregion
@@ -302,401 +253,6 @@ namespace Ryujinx.Common.Collections
             }
 
             return old;
-        }
-
-        /// <summary>
-        /// Returns the node with the largest key where <paramref name="node"/> is considered the root node.
-        /// </summary>
-        /// <param name="node">Root node</param>
-        /// <returns>Node with the maximum key in the tree of <paramref name="node"/></returns>
-        private static T Maximum(T node)
-        {
-            T tmp = node;
-            while (tmp.Right != null)
-            {
-                tmp = tmp.Right;
-            }
-
-            return tmp;
-        }
-
-        /// <summary>
-        /// Returns the node with the smallest key where <paramref name="node"/> is considered the root node.
-        /// </summary>
-        /// <param name="node">Root node</param>
-        /// <returns>Node with the minimum key in the tree of <paramref name="node"/></returns>
-        /// <exception cref="ArgumentNullException"><paramref name="node"/> is null</exception>
-        private static T Minimum(T node)
-        {
-            if (node == null)
-            {
-                throw new ArgumentNullException(nameof(node));
-            }
-            T tmp = node;
-            while (tmp.Left != null)
-            {
-                tmp = tmp.Left;
-            }
-
-            return tmp;
-        }
-
-        /// <summary>
-        /// Returns the node whose key immediately less than or equal to <paramref name="node"/>.
-        /// </summary>
-        /// <param name="node">Node for which to find the floor node of</param>
-        /// <returns>Node whose key is immediately less than or equal to <paramref name="node"/>, or null if no such node is found.</returns>
-        /// <exception cref="ArgumentNullException"><paramref name="node"/> is null</exception>
-        private T FloorNode(T node)
-        {
-            if (node == null)
-            {
-                throw new ArgumentNullException(nameof(node));
-            }
-            T tmp = Root;
-
-            while (tmp != null)
-            {
-                int cmp = node.CompareTo(tmp);
-                if (cmp > 0)
-                {
-                    if (tmp.Right != null)
-                    {
-                        tmp = tmp.Right;
-                    }
-                    else
-                    {
-                        return tmp;
-                    }
-                }
-                else if (cmp < 0)
-                {
-                    if (tmp.Left != null)
-                    {
-                        tmp = tmp.Left;
-                    }
-                    else
-                    {
-                        T parent = tmp.Parent;
-                        T ptr = tmp;
-                        while (parent != null && ptr == parent.Left)
-                        {
-                            ptr = parent;
-                            parent = parent.Parent;
-                        }
-                        return parent;
-                    }
-                }
-                else
-                {
-                    return tmp;
-                }
-            }
-            return null;
-        }
-
-        /// <summary>
-        /// Returns the node whose key is immediately greater than or equal to than <paramref name="node"/>.
-        /// </summary>
-        /// <param name="node">Node for which to find the ceiling node of</param>
-        /// <returns>Node whose key is immediately greater than or equal to <paramref name="node"/>, or null if no such node is found.</returns>
-        /// <exception cref="ArgumentNullException"><paramref name="node"/> is null</exception>
-        private T CeilingNode(T node)
-        {
-            if (node == null)
-            {
-                throw new ArgumentNullException(nameof(node));
-            }
-            T tmp = Root;
-
-            while (tmp != null)
-            {
-                int cmp = node.CompareTo(tmp);
-                if (cmp < 0)
-                {
-                    if (tmp.Left != null)
-                    {
-                        tmp = tmp.Left;
-                    }
-                    else
-                    {
-                        return tmp;
-                    }
-                }
-                else if (cmp > 0)
-                {
-                    if (tmp.Right != null)
-                    {
-                        tmp = tmp.Right;
-                    }
-                    else
-                    {
-                        T parent = tmp.Parent;
-                        T ptr = tmp;
-                        while (parent != null && ptr == parent.Right)
-                        {
-                            ptr = parent;
-                            parent = parent.Parent;
-                        }
-                        return parent;
-                    }
-                }
-                else
-                {
-                    return tmp;
-                }
-            }
-            return null;
-        }
-
-        #endregion
-
-        #region Internal Methods (RBL)
-
-        internal void RestoreBalanceAfterRemoval(T balanceNode)
-        {
-            T ptr = balanceNode;
-
-            while (ptr != Root && ColorOf(ptr) == Black)
-            {
-                if (ptr == LeftOf(ParentOf(ptr)))
-                {
-                    T sibling = RightOf(ParentOf(ptr));
-
-                    if (ColorOf(sibling) == Red)
-                    {
-                        SetColor(sibling, Black);
-                        SetColor(ParentOf(ptr), Red);
-                        RotateLeft(ParentOf(ptr));
-                        sibling = RightOf(ParentOf(ptr));
-                    }
-                    if (ColorOf(LeftOf(sibling)) == Black && ColorOf(RightOf(sibling)) == Black)
-                    {
-                        SetColor(sibling, Red);
-                        ptr = ParentOf(ptr);
-                    }
-                    else
-                    {
-                        if (ColorOf(RightOf(sibling)) == Black)
-                        {
-                            SetColor(LeftOf(sibling), Black);
-                            SetColor(sibling, Red);
-                            RotateRight(sibling);
-                            sibling = RightOf(ParentOf(ptr));
-                        }
-                        SetColor(sibling, ColorOf(ParentOf(ptr)));
-                        SetColor(ParentOf(ptr), Black);
-                        SetColor(RightOf(sibling), Black);
-                        RotateLeft(ParentOf(ptr));
-                        ptr = Root;
-                    }
-                }
-                else
-                {
-                    T sibling = LeftOf(ParentOf(ptr));
-
-                    if (ColorOf(sibling) == Red)
-                    {
-                        SetColor(sibling, Black);
-                        SetColor(ParentOf(ptr), Red);
-                        RotateRight(ParentOf(ptr));
-                        sibling = LeftOf(ParentOf(ptr));
-                    }
-                    if (ColorOf(RightOf(sibling)) == Black && ColorOf(LeftOf(sibling)) == Black)
-                    {
-                        SetColor(sibling, Red);
-                        ptr = ParentOf(ptr);
-                    }
-                    else
-                    {
-                        if (ColorOf(LeftOf(sibling)) == Black)
-                        {
-                            SetColor(RightOf(sibling), Black);
-                            SetColor(sibling, Red);
-                            RotateLeft(sibling);
-                            sibling = LeftOf(ParentOf(ptr));
-                        }
-                        SetColor(sibling, ColorOf(ParentOf(ptr)));
-                        SetColor(ParentOf(ptr), Black);
-                        SetColor(LeftOf(sibling), Black);
-                        RotateRight(ParentOf(ptr));
-                        ptr = Root;
-                    }
-                }
-            }
-            SetColor(ptr, Black);
-        }
-
-        internal void RestoreBalanceAfterInsertion(T balanceNode)
-        {
-            SetColor(balanceNode, Red);
-            while (balanceNode != null && balanceNode != Root && ColorOf(ParentOf(balanceNode)) == Red)
-            {
-                if (ParentOf(balanceNode) == LeftOf(ParentOf(ParentOf(balanceNode))))
-                {
-                    T sibling = RightOf(ParentOf(ParentOf(balanceNode)));
-
-                    if (ColorOf(sibling) == Red)
-                    {
-                        SetColor(ParentOf(balanceNode), Black);
-                        SetColor(sibling, Black);
-                        SetColor(ParentOf(ParentOf(balanceNode)), Red);
-                        balanceNode = ParentOf(ParentOf(balanceNode));
-                    }
-                    else
-                    {
-                        if (balanceNode == RightOf(ParentOf(balanceNode)))
-                        {
-                            balanceNode = ParentOf(balanceNode);
-                            RotateLeft(balanceNode);
-                        }
-                        SetColor(ParentOf(balanceNode), Black);
-                        SetColor(ParentOf(ParentOf(balanceNode)), Red);
-                        RotateRight(ParentOf(ParentOf(balanceNode)));
-                    }
-                }
-                else
-                {
-                    T sibling = LeftOf(ParentOf(ParentOf(balanceNode)));
-
-                    if (ColorOf(sibling) == Red)
-                    {
-                        SetColor(ParentOf(balanceNode), Black);
-                        SetColor(sibling, Black);
-                        SetColor(ParentOf(ParentOf(balanceNode)), Red);
-                        balanceNode = ParentOf(ParentOf(balanceNode));
-                    }
-                    else
-                    {
-                        if (balanceNode == LeftOf(ParentOf(balanceNode)))
-                        {
-                            balanceNode = ParentOf(balanceNode);
-                            RotateRight(balanceNode);
-                        }
-                        SetColor(ParentOf(balanceNode), Black);
-                        SetColor(ParentOf(ParentOf(balanceNode)), Red);
-                        RotateLeft(ParentOf(ParentOf(balanceNode)));
-                    }
-                }
-            }
-            SetColor(Root, Black);
-        }
-
-        internal void RotateLeft(T node)
-        {
-            if (node != null)
-            {
-                T right = RightOf(node);
-                node.Right = LeftOf(right);
-                if (LeftOf(right) != null)
-                {
-                    LeftOf(right).Parent = node;
-                }
-                right.Parent = ParentOf(node);
-                if (ParentOf(node) == null)
-                {
-                    Root = right;
-                }
-                else if (node == LeftOf(ParentOf(node)))
-                {
-                    ParentOf(node).Left = right;
-                }
-                else
-                {
-                    ParentOf(node).Right = right;
-                }
-                right.Left = node;
-                node.Parent = right;
-            }
-        }
-
-        internal void RotateRight(T node)
-        {
-            if (node != null)
-            {
-                T left = LeftOf(node);
-                node.Left = RightOf(left);
-                if (RightOf(left) != null)
-                {
-                    RightOf(left).Parent = node;
-                }
-                left.Parent = node.Parent;
-                if (ParentOf(node) == null)
-                {
-                    Root = left;
-                }
-                else if (node == RightOf(ParentOf(node)))
-                {
-                    ParentOf(node).Right = left;
-                }
-                else
-                {
-                    ParentOf(node).Left = left;
-                }
-                left.Right = node;
-                node.Parent = left;
-            }
-        }
-
-        #endregion
-
-        #region Safety-Methods
-
-        // These methods save memory by allowing us to forego sentinel nil nodes, as well as serve as protection against NullReferenceExceptions.
-
-        /// <summary>
-        /// Returns the color of <paramref name="node"/>, or Black if it is null.
-        /// </summary>
-        /// <param name="node">Node</param>
-        /// <returns>The boolean color of <paramref name="node"/>, or black if null</returns>
-        internal static bool ColorOf(T node)
-        {
-            return node == null || node.Color;
-        }
-
-        /// <summary>
-        /// Sets the color of <paramref name="node"/> node to <paramref name="color"/>.
-        /// <br></br>
-        /// This method does nothing if <paramref name="node"/> is null.
-        /// </summary>
-        /// <param name="node">Node to set the color of</param>
-        /// <param name="color">Color (Boolean)</param>
-        internal static void SetColor(T node, bool color)
-        {
-            if (node != null)
-            {
-                node.Color = color;
-            }
-        }
-
-        /// <summary>
-        /// This method returns the left node of <paramref name="node"/>, or null if <paramref name="node"/> is null.
-        /// </summary>
-        /// <param name="node">Node to retrieve the left child from</param>
-        /// <returns>Left child of <paramref name="node"/></returns>
-        internal static T LeftOf(T node)
-        {
-            return node?.Left;
-        }
-
-        /// <summary>
-        /// This method returns the right node of <paramref name="node"/>, or null if <paramref name="node"/> is null.
-        /// </summary>
-        /// <param name="node">Node to retrieve the right child from</param>
-        /// <returns>Right child of <paramref name="node"/></returns>
-        internal static T RightOf(T node)
-        {
-            return node?.Right;
-        }
-
-        /// <summary>
-        /// Returns the parent node of <paramref name="node"/>, or null if <paramref name="node"/> is null.
-        /// </summary>
-        /// <param name="node">Node to retrieve the parent from</param>
-        /// <returns>Parent of <paramref name="node"/></returns>
-        internal static T ParentOf(T node)
-        {
-            return node?.Parent;
         }
 
         #endregion
